Remove RayTransformConfiguration and move RayLauncherConfiguration
