+++ data-processing-lib/test/data_processing_tests/launch/pure_python/test_noop_launch.py
@@ -14,10 +14,7 @@ import os
 
 import pyarrow as pa
 
-from data_processing.launch.pure_python import PythonTransformLauncher
-from data_processing.launch.ray import RayLauncherConfiguration, RayTransformLauncher
 from data_processing.test_support.launch.transform_test import AbstractTransformLauncherTest
-from data_processing.test_support.transform import NOOPTransformConfiguration
 
 table = pa.Table.from_pydict({"name": pa.array(["Tom"]), "age": pa.array([23])})
 expected_table = table  # We're a noop after all.
