+++ data-processing-lib/test/data_processing_tests/launch/ray/launcher_test.py
@@ -15,8 +15,6 @@ import sys
 
 from data_processing.transform import TransformConfiguration
 from data_processing.launch.ray import (
-    DefaultTableTransformRuntimeRay,
-    RayLauncherConfiguration,
     RayTransformLauncher,
 )
 from data_processing.transform import AbstractTableTransform
