+++ transforms/code/code_quality/src/code_quality_transform.py
@@ -25,8 +25,7 @@ import numpy as np
 import pyarrow as pa
 from bs4 import BeautifulSoup
 from data_processing.transform import TransformConfiguration
-from data_processing.launch.pure_python import PythonTransformLauncher, PythonLauncherConfiguration
-from data_processing.launch.ray import RayLauncherConfiguration, RayTransformLauncher
+from data_processing.launch.ray import RayTransformLauncher
 from data_processing.transform import AbstractTableTransform
 from data_processing.utils import TransformUtils
 from transformers import AutoTokenizer
