+++ transforms/code/malware/src/malware_transform.py
@@ -20,7 +20,7 @@ import clamd
 import pyarrow as pa
 from data_processing.transform import TransformConfiguration
 from data_processing.launch.pure_python import PythonTransformLauncher, PythonLauncherConfiguration
-from data_processing.launch.ray import RayLauncherConfiguration, RayTransformLauncher
+from data_processing.launch.ray import RayTransformLauncher
 from data_processing.transform import AbstractTableTransform
 from data_processing.utils import get_logger
 from data_processing.utils.transform_utils import TransformUtils
