+++ transforms/universal/filter/src/filter_transform.py
@@ -18,7 +18,7 @@ import duckdb
 import pyarrow as pa
 from data_processing.transform import TransformConfiguration
 from data_processing.launch.pure_python import PythonTransformLauncher, PythonLauncherConfiguration
-from data_processing.launch.ray import RayLauncherConfiguration, RayTransformLauncher
+from data_processing.launch.ray import RayTransformLauncher
 from data_processing.transform import AbstractTableTransform
 from data_processing.utils import CLIArgumentProvider, get_logger
